src/__test__/__test__/news.test.ts
import request from 'supertest';
import express, { Express } from 'express';
import newsRoutes from '../../routes/news'; // Replace with the actual path
import {faker} from 'faker';

const app: Express = express();
app.use(express.json());
app.use('/', newsRoutes);

describe('News Routes', () => {
  // Test the GET route for fetching all news
  it('should return all news', async () => {
    const response = await request(app).get('/api/news/');
    expect(response.status).toBe(200);
    // Add assertions based on generated fake data using faker
  });

  // Test the GET route for news by preferences
  it('should return news by user preferences', async () => {
    const userId = faker.random.uuid();
    const response = await request(app).get(`/api/news/userprefs?userId=${userId}`);
    expect(response.status).toBe(200);
    // Add assertions based on generated fake data using faker
  });

  // Test the GET route for searching news by keyword
  it('should return news by keyword search', async () => {
    const keyword = faker.lorem.word();
    const response = await request(app).get(`/api/news/search/${keyword}`);
    expect(response.status).toBe(200);
    // Add assertions based on generated fake data using faker
  });

  // Test routes for other scenarios and edge cases
  it('should handle user not found when fetching user preferences', async () => {
    const nonexistentUser = faker.random.uuid();
    const response = await request(app).get(`/api/news/userprefs?userId=${nonexistentUser}`);
    expect(response.status).toBe(404);
    // Add assertions based on generated fake data using faker
  });

  // Add more test cases as needed
});
